refactor: Move string and set function tests to separate files
Extract inline test modules from string_functions.rs and set_functions.rs
into dedicated test files under src/vm/tests/. Maintains all 149 tests
(101 string + 48 set) with identical behavior.
