+++ src/vm/tests/mod.rs
@@ -1,3 +1,5 @@
 mod basic;
 mod math_variadic;
 mod math_errors;
+mod string_functions;
+mod set_functions;
